+++ src/pages/CoreFeatures/analytics/AnalyticsDashboard.tsx
@@ -4,12 +4,22 @@ import { Line, Bar, Pie } from 'react-chartjs-2';
 import { ClipLoader } from 'react-spinners';
 import { saveAs } from 'file-saver';
 import { json2csv } from 'json-2-csv';
-import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, BarElement, ArcElement, Tooltip, Legend } from 'chart.js';
+import {
+  Chart as ChartJS,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  LineElement,
+  BarElement,
+  ArcElement,
+  Tooltip,
+  Legend,
+} from 'chart.js';
 import annotationPlugin from 'chartjs-plugin-annotation';
-import moment from 'moment';
-import Navbar from '../components/Navbar';
-import Footer from '../components/Footer';
-import { dummyCampaigns, generateRealTimeData } from '../utils/dummyData';
+import dayjs from 'dayjs';
+import Navbar from '../../../components/Navbar';
+import Footer from '../../../components/Footer';
+import { dummyCampaigns, generateRealTimeData } from '../../../utils/dummyData';
 
 const { RangePicker } = DatePicker;
 const { Option } = Select;
@@ -47,30 +57,37 @@ const AnalyticsDashboard: React.FC = () => {
   const [campaigns, setCampaigns] = useState<Campaign[]>([]);
   const [selectedCampaignId, setSelectedCampaignId] = useState<string | null>(null);
   const [darkMode, setDarkMode] = useState(false);
-  const [dateRange, setDateRange] = useState<[moment.Moment, moment.Moment] | null>(null);
+  const [dateRange, setDateRange] = useState<[dayjs.Dayjs, dayjs.Dayjs] | null>(null);
   const [metricType, setMetricType] = useState('Open Rate');
 
   useEffect(() => {
-    setCampaigns(dummyCampaigns);
-    setSelectedCampaignId(dummyCampaigns[0]?.id || null);
-  }, []);
+  // Set initial campaigns and selected campaign
+  setCampaigns(dummyCampaigns);
+  if (dummyCampaigns.length > 0) {
+    setSelectedCampaignId(dummyCampaigns[0].id);
+  }
+}, [dummyCampaigns]);
 
-  useEffect(() => {
-    if (!selectedCampaignId) return;
+useEffect(() => {
+  if (!selectedCampaignId) return;
 
-    const interval = setInterval(() => {
-      const updatedMetrics = generateRealTimeData(selectedCampaignId);
-      setCampaigns((prev) =>
-        prev.map((campaign) =>
-          campaign.id === selectedCampaignId
-            ? { ...campaign, metrics: updatedMetrics }
-            : campaign
-        )
-      );
-    }, 5000);
+  const interval = setInterval(() => {
+    const updatedMetrics = generateRealTimeData(selectedCampaignId);
 
-    return () => clearInterval(interval);
-  }, [selectedCampaignId]);
+    // Update only if there is a change in metrics
+    setCampaigns((prevCampaigns) =>
+      prevCampaigns.map((campaign) =>
+        campaign.id === selectedCampaignId &&
+        JSON.stringify(campaign.metrics) !== JSON.stringify(updatedMetrics)
+          ? { ...campaign, metrics: updatedMetrics }
+          : campaign
+      )
+    );
+  }, 5000);
+
+  // Cleanup interval on component unmount or dependency change
+  return () => clearInterval(interval);
+}, [selectedCampaignId]);
 
   const exportToCSV = async () => {
     try {
@@ -78,6 +95,7 @@ const AnalyticsDashboard: React.FC = () => {
       const blob = new Blob([csvData], { type: 'text/csv;charset=utf-8;' });
       saveAs(blob, 'analytics-data.csv');
     } catch (error) {
+      alert('Failed to export CSV. Please try again.');
       console.error('Error exporting CSV:', error);
     }
   };
@@ -86,7 +104,11 @@ const AnalyticsDashboard: React.FC = () => {
 
   if (!campaigns.length) {
     return (
-      <div className={`min-h-screen flex items-center justify-center ${darkMode ? 'bg-gray-900 text-yellow-500' : 'bg-white text-black'}`}>
+      <div
+        className={`min-h-screen flex items-center justify-center ${
+          darkMode ? 'bg-gray-900 text-yellow-500' : 'bg-white text-black'
+        }`}
+      >
         <ClipLoader color={darkMode ? '#fbbf24' : '#1f2937'} size={50} />
         <p className="mt-4 text-lg font-semibold">Loading...</p>
       </div>
@@ -103,9 +125,9 @@ const AnalyticsDashboard: React.FC = () => {
 
   const { metrics, dailyEngagement } = selectedCampaign;
 
-  const filteredEngagement = dateRange?.[0] && dateRange?.[1]
+  const filteredEngagement = dateRange
     ? dailyEngagement.filter((entry) =>
-        moment(entry.date).isBetween(dateRange[0], dateRange[1], null, '[]')
+        dayjs(entry.date).isBetween(dateRange[0], dateRange[1], null, '[]')
       )
     : dailyEngagement;
 
@@ -115,7 +137,7 @@ const AnalyticsDashboard: React.FC = () => {
     'Response Rate': 'responseRate',
   };
 
-  const selectedMetricKey = metricKeyMap[metricType];
+  const selectedMetricKey = metricKeyMap[metricType as keyof typeof metricKeyMap];
 
   const lineData = {
     labels: filteredEngagement.map((entry) => entry.date),
